src/requirements.txt
anyio==3.6.1; python_full_version >= "3.6.2" and python_version >= "3.7"
cffi==1.15.1; python_version >= "3.6"
click==8.1.3; python_version >= "3.7"
colorama==0.4.5; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.5.0" and platform_system == "Windows"
cryptography==38.0.2; python_version >= "3.6"
ecdsa==0.18.0; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
fastapi==0.85.0; python_version >= "3.7"
gunicorn==20.1.0; python_version >= "3.5"
h11==0.14.0; python_version >= "3.7"
httptools==0.5.0; python_version >= "3.7" and python_full_version >= "3.5.0"
idna==3.4; python_full_version >= "3.6.2" and python_version >= "3.7"
mangum==0.15.1; python_version >= "3.7"
numpy==1.23.4
pandas==1.5.0; python_version >= "3.8"
pyasn1==0.4.8; python_version >= "3.6" and python_version < "4"
pycparser==2.21; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.4.0"
pydantic==1.10.2; python_version >= "3.7"
pyotp==2.7.0; python_version >= "3.6"
python-dateutil==2.8.2; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
python-dotenv==0.21.0; python_version >= "3.7"
python-jose==3.3.0
pytz==2022.4; python_version >= "3.8"
pyyaml==6.0; python_version >= "3.7"
rsa==4.9; python_version >= "3.6" and python_version < "4"
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
sniffio==1.3.0; python_full_version >= "3.6.2" and python_version >= "3.7"
starlette==0.20.4; python_version >= "3.7"
typing-extensions==4.4.0; python_version < "3.10" and python_version >= "3.7"
uvicorn==0.18.3; python_version >= "3.7"
uvloop==0.17.0; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.7"
watchfiles==0.17.0; python_version >= "3.7"
websockets==10.3; python_version >= "3.7"
